lab2/requirements.txt
Pillow==6.1.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
matplotlib==3.1.1msgpack-numpy==0.4.4.3
numpy==1.16.4
numpydoc==0.9.1
opencv-python==4.2.0.34
